correcion de error en dato de la columna dpi tabla solicitudes

+++ database/migrations/2024_05_10_203426_create_solicitudes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nombre');
             $table->string('apellido');
-            $table->integer('dpi')->unique();
+            $table->bigInteger('dpi')->unique();
             $table->integer('telefono');
             $table->string('direccion');
             $table->float('ingresos');
